backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
alembic==1.13.1
psycopg2-binary==2.9.9
pillow==10.2.0
torch==2.1.0
torchvision==0.16.0
transformers==4.36.0
diffusers==0.25.0
accelerate>=0.25.0
safetensors>=0.4.1
opencv-python>=4.8.0
numpy>=1.24.0
redis>=5.0.0
boto3>=1.34.0
huggingface-hub>=0.20.0
pytest==8.0.1
httpx==0.26.0
i18next==0.0.3
langdetect==1.0.9 